Implement St8ger
This commit contains the implementation of the `cube st8ger` and `cube recipe`. The `recipe` package is the code running in a container when an OPI <staging> task is desired. It is basically a wrapper around `packs` and extends it with the app-bits download and droplet upload functionality. In addition this code extends the `k8s` package with a `task desirer`, which translates an `opi.Task` to a Kubernetes job.